Update to Main file- log file path
When running free climber on a macbook air and python 3.7, the path of the log folder was concatenated with the name of the parent directory leading to an error when the program tried to find the correct folder to log progress. By changing the lines of code above this error was avoided and the program ran without errors.